Refactored and moved specific patches to /Patches. Moved adding new abilities, cards, regions to /Models
